move hello to separate file
