+++ crates/net/eth-wire/src/lib.rs
@@ -13,6 +13,7 @@ pub mod error;
 mod ethstream;
 mod p2pstream;
 mod pinger;
+mod hello;
 pub use builder::*;
 pub mod types;
 pub use types::*;
@@ -24,6 +25,7 @@ pub use tokio_util::codec::{
 
 pub use crate::{
     disconnect::DisconnectReason,
+    hello::HelloMessage,
     ethstream::{EthStream, UnauthedEthStream, MAX_MESSAGE_SIZE},
-    p2pstream::{HelloMessage, P2PStream, ProtocolVersion, UnauthedP2PStream},
+    p2pstream::{P2PStream, ProtocolVersion, UnauthedP2PStream},
 };
